Replace ByteBuffer use with Buffer and SingleBuffer

+++ src/main/java/com/maxmind/db/CtrlData.java
@@ -1,4 +1,4 @@
 package com.maxmind.db;
 
-record CtrlData(Type type, int ctrlByte, int offset, int size) {
+record CtrlData(Type type, int ctrlByte, long offset, int size) {
 }
